% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundleCompile.R
\name{bundleCompile}
\alias{bundleCompile}
\title{bundleCompile}
\usage{
bundleCompile(bundle, regionmapping = getConfig("regionmapping"), ...)
}
\arguments{
\item{bundle}{path to a bundle file}

\item{regionmapping}{region mapping to be used for aggregation. If not
specified the current default will be used.}

\item{...}{(Optional) Settings that should be changed in addition. NOTE:
which settings can be modified varies from bundle to bundle and unsupported
choices can lead to errors.}
}
\description{
Function which takes a bundle tgz-file as create via \code{\link{retrieveData}}
and compiles the corresponding output with the given settings (in most cases
with a given region mapping)
}
\examples{
\dontrun{
bundleCompile("rev1_bundle_example.tgz", regionmapping = "regionmappingH12.csv")
}
}
\seealso{
\code{\link{retrieveData}},\code{\link{setConfig}}
}
\author{
Jan Philipp Dietrich
}
