% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheCopy.R
\name{cacheCopy}
\alias{cacheCopy}
\title{cacheCopy}
\usage{
cacheCopy(file, target = NULL, filter = NULL)
}
\arguments{
\item{file}{Path to a log file or content of a log as character vector.}

\item{target}{Folder to which the files should be copied. If NULL no data is copied.}

\item{filter}{Regular expression to filter the cache files shown in the log file.}
}
\value{
A vector of cache files which match the given log information and filter.
}
\description{
Copy cache files which were used for a given preprocessing.
}
\seealso{
Other cache management:
\code{\link[=cacheCleanup]{cacheCleanup()}},
\code{\link[=registerCacheFormat]{registerCacheFormat()}}
}
\author{
Jan Philipp Dietrich
}
\concept{cache management}
