% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cachePut.R
\name{cachePut}
\alias{cachePut}
\title{cachePut}
\usage{
cachePut(x, prefix, type, fname, callString)
}
\arguments{
\item{x}{data that should be written to cache}

\item{prefix}{function prefix (e.g. "calc" or "read")}

\item{type}{output type (e.g. "TauTotal")}

\item{fname}{The name of the cache file to be written.}

\item{callString}{A string representation of the function call that leads
to the cache file being written. Will be attached as an attribute.}
}
\value{
\code{fname} if the cache file was written, otherwise NULL. Writing a
cache file is optional and allowed to fail, so callers which need to know
whether the file exists (e.g. to list it in a puc file) must use this return
value rather than assume \code{fname} was created.
}
\description{
Save data to cache
}
\examples{
\dontrun{
fname <- madrat:::cacheNames("calc", "Example")$write
madrat:::cachePut(1, "calc", "Example", fname, 'calcOutput("Example")')
}
}
\seealso{
\code{\link{cachePut}}, \code{\link{cacheNames}}
}
\author{
Jan Philipp Dietrich, Pascal Sauer
}
\keyword{internal}
