% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanComment.R
\name{cleanComment}
\alias{cleanComment}
\title{cleanComment}
\usage{
cleanComment(
  x,
  remove = c("unit", "description", "comment", "origin", "creation date", "note")
)
}
\arguments{
\item{x}{magclass object the comment should be read from}

\item{remove}{Vector of categories to be removed}
}
\description{
Helper function to clean a comment from additional metadata information
}
\examples{
x <- maxample("animal")
getComment(x) <- c("unit: bla", "comment: hallo", "blub: ble")
madrat:::cleanComment(x)
}
\author{
Jan Philipp Dietrich
}
