% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commandAvailable.R
\name{commandAvailable}
\alias{commandAvailable}
\title{commandAvailable}
\usage{
commandAvailable(command)
}
\arguments{
\item{command}{The command to check, e.g. ls.}
}
\value{
TRUE or FALSE, depending on whether the command is available.
}
\description{
Check if a command line tool is available by trying to run `which <command>`.
}
