% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMadratOutputs.R
\name{compareMadratOutputs}
\alias{compareMadratOutputs}
\title{Compare a madrat function's output with and without your changes}
\usage{
compareMadratOutputs(package, functionName, subtypes)
}
\arguments{
\item{package}{[character(1)] The package where the given function is located.}

\item{functionName}{The name of the function from which you want to compare outputs.}

\item{subtypes}{The subtypes you want to check. For calc functions this must be NULL.}
}
\description{
With `compareMadratOutputs` you can easily compare the output of a madrat function (read, calc, ...) with and without
your changes. First, run `compareMadratOutputs` without your changes, so a `.rds` file with the original output is
saved. Then apply your changes and run `compareMadratOutputs` again to compare the new output to the original output.
}
\details{
For comparing `waldo::compare` is used if available, otherwise `all.equal`. If there are differences a `*-new.rds`
containing the new output is saved for closer inspection. All files are created in the current working directory.
}
\examples{
\dontrun{
# save readFRA2020-old.rds
compareMadratOutputs("mrcommons", "readFRA2020", c("forest_area", "deforestation")

# now apply your changes to mrcommons:::readFRA2020, reinstall mrcommons, restart the R session

# compare to readFRA2020-old.rds
compareMadratOutputs("mrcommons", "readFRA2020", c("forest_area", "deforestation")
}

}
\author{
Pascal Führlich
}
