% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTau.R
\name{convertTau}
\alias{convertTau}
\title{Convert Tau}
\usage{
convertTau(x)
}
\arguments{
\item{x}{MAgPIE object containing tau values and corresponding weights xref
at 0.5deg cellular level.}
}
\value{
Tau data and weights as MAgPIE object aggregated to country level
}
\description{
Convert landuse intensity data (tau) to data on ISO country level.
}
\author{
Jan Philipp Dietrich
}
