% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosticSources.R
\name{diagnosticSources}
\alias{diagnosticSources}
\title{diagnosticSources}
\usage{
diagnosticSources(file)
}
\arguments{
\item{file}{path to a diagnostics.csv file (as created by retrieveData)}
}
\value{
a vector with sources detected in the diagnostics.csv
}
\description{
Function to extract sources from a diagnostics.csv file
}
\seealso{
\code{\link{retrieveData}}
}
\author{
Jan Philipp Dietrich
}
