% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadSource.R
\name{downloadSource}
\alias{downloadSource}
\title{downloadSource}
\usage{
downloadSource(type, overwrite = FALSE)
}
\arguments{
\item{type}{source type, e.g. "IEA". A list of all available source types
can be retrieved with function \code{\link{getSources}("download")}.}

\item{overwrite}{Boolean deciding whether existing data should be
overwritten or not.}
}
\description{
Download a source. The function is a wrapper for specific functions designed
for the different possible source types.
}
\examples{

\dontrun{ 
a <- downloadSource("Tau")
}

}
\seealso{
\code{\link{setConfig}}, \code{\link{readSource}}
}
\author{
Jan Philipp Dietrich
}
