% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file2destination.R
\name{file2destination}
\alias{file2destination}
\title{Tool: file2destination}
\usage{
file2destination(file, destination)
}
\arguments{
\item{file}{The name of the file which should be mapped to a destination}

\item{destination}{The path relative to the main folder of the model to
which the file should be copied. In the case that the file should be copied
to more than one destination within the model data should be provided as a
vector of destinations.}
}
\description{
Function which maps a file of the data output folder to the
destination path in the corresponding folder. Mapping is stored
in a the data output folder in a file called 'file2destination.txt'
}
\examples{

\dontrun{
file2destination("example.txt","example/folder")
}
}
\seealso{
\code{\link{calcOutput}}, \code{\link{retrieveData}}
}
\author{
Jan Philipp Dietrich
}
