% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprint.R
\name{fingerprint}
\alias{fingerprint}
\title{Tool: fingerprint}
\usage{
fingerprint(folder, ...)
}
\arguments{
\item{folder}{A folder containing objects for which the fingerprint should
be created (all files in that folder and all sub-folders will be considered)}

\item{...}{R data objects that should be considered for the fingerprint as
well (e.g. functions, variables,...)}
}
\value{
A md5-based fingerprint of all provided sources
}
\description{
Function which creates a fingerprint of a folder together with some R
objects (e.g. functions). Based on the fingerprint it is possible to decide
whether there were some changes in the given folder and the given objects or
not. If the fingerprint is unchanged also all files and objects stayed the
same, otherwise the fingerprint changes
}
\examples{

\dontrun{
fingerprint(".",ls,c)
}
}
\seealso{
\code{\link{readSource}}
}
\author{
Jan Philipp Dietrich
}
