% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullEXAMPLE.R
\name{fullEXAMPLE}
\alias{fullEXAMPLE}
\title{fullExample}
\usage{
fullEXAMPLE(rev = 0)
}
\arguments{
\item{rev}{data revision which should be used as input (positive numeric).
\code{\link{setConfig}} (e.g. for setting the mainfolder if not already set
properly).}
}
\description{
Example for class of fullX functions. Can be used as template for a new function
or for testing the basic functionality
}
\examples{

\dontrun{ 
retrieveData("Test",rev=12,regionmapping="regionmappingH12.csv")
}

}
\seealso{
\code{\link{readSource}},\code{\link{getCalculations}},\code{\link{calcOutput}},\code{\link{setConfig}}
,\code{\link{file2destination}}
}
\author{
Jan Philipp Dietrich
}
