% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionCallString.R
\name{functionCallString}
\alias{functionCallString}
\title{functionCallString}
\usage{
functionCallString(functionName, argumentValues)
}
\arguments{
\item{functionName}{name of the called function}

\item{argumentValues}{the list of arguments passed}
}
\value{
A string representing the given function call
}
\description{
Create a string representation for a function call. If the resulting string
is longer than getConfig("maxLengthLogMessage") arguments are printed as
passed (e.g. as variable name instead of the evaluated content of that
variable), if that is still too long it is cropped.
}
\author{
Pascal Sauer
}
