% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConfig.R
\name{getConfig}
\alias{getConfig}
\title{getConfig}
\usage{
getConfig(option = NULL, raw = FALSE, verbose = TRUE, print = FALSE)
}
\arguments{
\item{option}{The option for which the setting should be returned. If set to
NULL all options are returned.}

\item{raw}{If set to FALSE some settings will be calculated, e.g. if the
cache folder is set to FALSE the full path will be calculated using the
main folder, or if the verbosity is not set the default verbosity will be
returned. If raw is set to TRUE settings are returned as they are currently
stored.}

\item{verbose}{boolean deciding whether status information/updates should be shown or not}

\item{print}{if TRUE and verbose is TRUE a configuration overview will also get printed}
}
\value{
A config list with all settings currently set for the madrat package
}
\description{
This function returns the madrat config which is currently loaded. If no
configuration has been loaded so far the configuration will be initialized
with default settings or system settings (if available).
}
\seealso{
\code{\link{setConfig}}, \code{\link{initializeConfig}}
}
\author{
Jan Philipp Dietrich
}
