% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDependentCalls.R
\name{getDependentCalls}
\alias{getDependentCalls}
\title{getDependentCalls}
\usage{
getDependentCalls(type, ...)
}
\arguments{
\item{type}{Type for which the calls it depends on should be retrieved}

\item{...}{additional arguments passed on}
}
\value{
a list containing n lists depending on the number of Calculations available for \code{type}.
Each list contains the associated calls as character vectors
}
\description{
Function to retrieve all calls of calcOutput and redSource
inside a calcOutput or readSource call.
}
\examples{
getDependentCalls("FAO")
getDependentCalls("fullMAGPIE")
}
\author{
Stephen Wirth, Jan Philipp Dietrich
}
