% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFlags.R
\name{getFlags}
\alias{getFlags}
\title{getFlags}
\usage{
getFlags(code)
}
\arguments{
\item{code}{A character vector with code from functions to be analyzed}
}
\value{
A list of found flag entries
}
\description{
Support function which extracts flags from code. Flags are string literals
in a function body, for example `"!# @pucArguments extra"`.
}
\seealso{
\code{\link{getCode}}

Other dependency graph: 
\code{\link{fingerprint}()},
\code{\link{getCode}()},
\code{\link{getDependencies}()},
\code{\link{getMadratGraph}()},
\code{\link{getMadratInfo}()},
\code{\link{visualizeDependencies}()}
}
\author{
Jan Philipp Dietrich
}
\concept{dependency graph}
