% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getISOlist.R
\name{getISOlist}
\alias{getISOlist}
\title{get official ISO list}
\usage{
getISOlist(type = "all", threshold = 1)
}
\arguments{
\item{type}{Determines what countries should be returned. "all" returns all
countries, "important" returns all countries which are above the population
threshold set in the configuration and "dispensable" returns all countries
which are below the threshold.}

\item{threshold}{Population threshold in million capita which determines whether
the country is put into the "important" or "dispensable" class
(default = 1 mio. people)}
}
\value{
vector of default ISO country codes.
}
\description{
Function which returns the ISO list which is used as default for the input
data preparation. It contains the countries to which all source data has to
be up- or downscaled to.
}
\note{
Please always use this function instead of directly referring to the data
object as the format in this data list might change in the future!
}
\examples{

head(getISOlist())
head(getISOlist("dispensable"))
}
\seealso{
\code{\link{getSources}}, \code{\link{getCalculations}}
}
\author{
Jan Philipp Dietrich
}
