% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocation.R
\name{getLocation}
\alias{getLocation}
\title{getLocation}
\usage{
getLocation(name, packages = installedMadratUniverse(), globalenv = TRUE)
}
\arguments{
\item{name}{name of the function to be found. Can be either the full name
(e.g. "calcTauTotal"), or just the type name (e.g. "TauTotal").}

\item{packages}{A character vector with packages in which should be looked
for the function}

\item{globalenv}{Boolean deciding whether functions in the global environment
should be included or not}
}
\value{
vector of packages in which a function matching the description
could be found
}
\description{
Returns names of packages in which functions matching the description
are found
}
\seealso{
\code{\link{getCalculations}}, \code{\link{getDependencies}}
}
\author{
Jan Philipp Dietrich
}
