% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMadratGraph.R
\name{getMadratGraph}
\alias{getMadratGraph}
\title{getMadratGraph}
\usage{
getMadratGraph(
  packages = installedMadratUniverse(),
  globalenv = getConfig("globalenv")
)
}
\arguments{
\item{packages}{A character vector with packages for which the available Sources/Calculations should be returned}

\item{globalenv}{Boolean deciding whether sources/calculations in the global environment should be included or not}
}
\value{
A data frame with 4 columns: from (source function), from_package (package the source function originates
from), to (function which is using the source), to_package (package of the using function)
}
\description{
Function returns the madrat graph of all linkages of full, calc, and read functions of the given madrat
based packages. Linkages to subfunctions of read functions (i.e. download, correct or convert functions)
are not listed separately, but collectively referred to through the corresponding read function.
}
\seealso{
\code{\link{getCalculations}}, \code{\link{getConfig}}
}
\author{
Jan Philipp Dietrich
}
