% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMadratInfo.R
\name{getMadratInfo}
\alias{getMadratInfo}
\title{getMadratInfo}
\usage{
getMadratInfo(graph = NULL, cutoff = 5, extended = FALSE, ...)
}
\arguments{
\item{graph}{A madrat graph as returned by \code{\link{getMadratGraph}}.
Will be created with \code{\link{getMadratGraph}} if not provided.}

\item{cutoff}{Integer introducing a cutoff of items to be returned for
outputs which can become quite verbose.}

\item{extended}{Will add additional outputs which has been removed
from standard output due to limited usefulness.}

\item{...}{Additional arguments for \code{\link{getMadratGraph}} in case
that no graph is provided (otherwise ignored)}
}
\description{
Collects and returns detailed information about the currently loaded
network of madrat functions.
}
\seealso{
\code{\link{getCalculations}}, \code{\link{getMadratGraph}}
}
\author{
Jan Philipp Dietrich
}
