% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMainfolder.R
\name{getMainfolder}
\alias{getMainfolder}
\title{getMainfolder}
\usage{
getMainfolder(verbose = TRUE)
}
\arguments{
\item{verbose}{boolean deciding whether status information/updates should be shown or not}
}
\description{
Functions checks for a global setting of the mainfolder (either by setting the environment
variable "MADRAT_MAINFOLDER" or by setting the R option with the same name). If none of these
is available the user will be asked for a directory. If this is not provided a temporary folder
will be used.
}
\seealso{
\code{\link{initializeConfig}}, \code{\link{getConfig}}, \code{\link{setConfig}}
}
\author{
Jan Philipp Dietrich
}
