% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeConfig.R
\name{initializeConfig}
\alias{initializeConfig}
\title{initializeConfig}
\usage{
initializeConfig(verbose = TRUE)
}
\arguments{
\item{verbose}{boolean deciding whether status information/updates should be shown or not}
}
\description{
Checks whether configuration already has been set. If not, it will be initialized
with default settings or (if available) system settings.
}
\seealso{
\code{\link{getMainfolder}}, \code{\link{getConfig}}, \code{\link{setConfig}}
}
\author{
Jan Philipp Dietrich
}
