% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madrat-package.R
\docType{package}
\name{madrat-package}
\alias{madrat-package}
\alias{madrat}
\title{May All Data be Reproducible And Transparent (madrat) *}
\description{
Package provides a basic framework which should improve reproducibility and transparency in data processing. 
It provides functionality such as automatic meta data creation and management, rudimentary quality management, 
data caching, work-flow management an data aggregation.
}
\details{
* The title is a wish not a promise. By no means we expect this package to deliver everything what is needed to
achieve full reproducibility and transparency, but we believe that it supports efforts in this direction.
}
