% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madratAttach.R
\name{madratAttach}
\alias{madratAttach}
\alias{madratDetach}
\title{madratAttach / madratDetach}
\usage{
madratAttach(package)

madratDetach(package)
}
\arguments{
\item{package}{name of the package to be loaded. Alternative, the path to
the package.}
}
\description{
Attaches the madrat functions of a package to the currently active madrat
universe or detaches it again from it.
}
\section{Functions}{
\itemize{
\item \code{madratDetach()}: detach package from madrat universe

}}
\examples{
\dontrun{
madratAttach("madrat")
}
}
\seealso{
\code{\link{getConfig}}, \code{\link{setConfig}}
}
\author{
Jan Philipp Dietrich
}
