% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madTempDir.R
\name{madTempDir}
\alias{madTempDir}
\title{madTempDir}
\usage{
madTempDir()
}
\value{
path to the temp folder
}
\description{
returns a temporary directory as a subfolder of the tempfolder set in
\code{getConfig("tmpfolder")}.
}
\examples{
\dontrun{
madrat:::madTempDir()
}

}
\seealso{
\code{\link{getConfig}}

Other madrat configuration: 
\code{\link{getConfig}()},
\code{\link{getMainfolder}()},
\code{\link{initializeConfig}()},
\code{\link{madratAttach}()},
\code{\link{setConfig}()}
}
\author{
Jan Philipp Dietrich
}
\concept{madrat configuration}
