% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoryProfiling.R
\name{memoryProfilingSupported}
\alias{memoryProfilingSupported}
\title{memoryProfilingSupported}
\usage{
memoryProfilingSupported()
}
\value{
TRUE if memory profiling can be performed, FALSE otherwise
}
\description{
Checks whether the kernel interfaces required for memory profiling are available.
Only Linux exposes both the high-water mark of the resident set size and a way
to reset it.
}
\seealso{
\code{\link{setConfig}}
}
\author{
Patrick Rein
}
\keyword{internal}
