% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadataGFZ.R
\name{metadataGFZ}
\alias{metadataGFZ}
\title{metadataGFZ}
\usage{
metadataGFZ(doi)
}
\arguments{
\item{doi}{DOI of a data set hosted at GFZ dataservices}
}
\value{
a list with entries "license", "citation", "authors" and "year"
}
\description{
Function to extract metadata information of a data set hosted at GFZ dataservices
(https://dataservices.gfz-potsdam.de/portal/).
}
\examples{
\dontrun{
metadataGFZ("10.5880/pik.2019.004")
}
}
\seealso{
\code{\link{toolstartmessage}}, \code{\link{vcat}}
}
\author{
Jan Philipp Dietrich
}
