% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepComment.R
\name{prepComment}
\alias{prepComment}
\title{prepComment}
\usage{
prepComment(x, name, warning = NULL)
}
\arguments{
\item{x}{content to be added as metadata comment}

\item{name}{Name of the metadata entry}

\item{warning}{Either NULL (no warning) or a warning text that should be
returned if x is NULL}
}
\description{
Helper function to properly format a metadata comment entry
}
\examples{
 madrat:::prepComment("example comment", "example")
}
\author{
Jan Philipp Dietrich
}
