% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepExtendedComment.R
\name{prepExtendedComment}
\alias{prepExtendedComment}
\title{prepExtendedComment}
\usage{
prepExtendedComment(x, type = "#undefined", warn = TRUE, n = 1)
}
\arguments{
\item{x}{list containing the metadata to be condensed}

\item{type}{output type, e.g. "TauTotal"}

\item{warn}{boolean indicating whether warnings should be triggered
if entries are missing, or not.}

\item{n}{the number of functions to go back for the extraction of the call
information}
}
\description{
Helper function condense metadata information into
an extended comment entry
}
\examples{
test <- function(a = 1) {
  return(madrat:::prepExtendedComment(list(unit = "m", description = "example", package = "blub")))
}
test(a = 42)

}
\author{
Jan Philipp Dietrich
}
