% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepFunctionName.R
\name{prepFunctionName}
\alias{prepFunctionName}
\title{prepFunctionName}
\usage{
prepFunctionName(type, prefix = "calc", ignore = NULL, error_on_missing = TRUE)
}
\arguments{
\item{type}{name of calculation/source}

\item{prefix}{Type of calculations. Available options are "download" (source download),
"read" (source read), "correct" (source corrections), "convert" (source conversion to ISO countries),
"calc" (further calculations), and "full" (collections of calculations)}

\item{ignore}{vector of arguments which should be ignored (not be part of the function call)}

\item{error_on_missing}{boolean deciding whether a missing type should throw an error or return NULL}
}
\value{
A function call as character to the specified function with corresponding package as attribute
}
\description{
Function to prepare a function call for a given type and prefix
}
\examples{

print(madrat:::prepFunctionName("Tau", "read"))
print(madrat:::prepFunctionName("TauTotal", "calc"))
print(madrat:::prepFunctionName("EXAMPLE", "full"))
}
\seealso{
\code{\link{readSource}}, \code{\link{setConfig}}
}
\author{
Jan Philipp Dietrich
}
