% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSource.R
\name{readSource}
\alias{readSource}
\title{readSource}
\usage{
readSource(type, subtype = NULL, subset = NULL, convert = TRUE)
}
\arguments{
\item{type}{source type, e.g. "IEA". A list of all available source types
can be retrieved with function \code{\link{getSources}}.}

\item{subtype}{For some sources there are subtypes of the source, for these
source the subtype can be specified with this argument. If a source does not
have subtypes, subtypes should not be set.}

\item{subset}{Similar to \code{subtype} a source can also have \code{subsets}. A \code{subsets}
can be used to only read part of the data. This can in particular make sense for huge
data sets where reading in the whole data set might be impractical or even
infeasible.}

\item{convert}{Boolean indicating whether input data conversion to
ISO countries should be done or not. In addition it can be set to "onlycorrect"
for sources with a separate correctXXX-function.}
}
\value{
Either a magpie object, or a list with the entries x (the data object)
and class (the class of x). The temporal and data dimensionality
should match the source data. The spatial dimension should either match the source data or,
if the convert argument is set to TRUE, should be on ISO code country level. For magpie objects
magclass::clean_magpie is run and if convert = TRUE ISO code country level is checked.
}
\description{
Read in a source file and convert it to a MAgPIE object. The function is a
wrapper for specific functions designed for the different possible source
types.
}
\examples{
\dontrun{
a <- readSource("Tau", "paper")
}

}
\seealso{
\code{\link{setConfig}}, \code{\link{downloadSource}}, \code{\link{readTau}}
}
\author{
Jan Philipp Dietrich, Anastasis Giannousakis, Lavinia Baumstark, Pascal Führlich
}
