% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redirectTau.R
\name{redirectTau}
\alias{redirectTau}
\title{redirectTau}
\usage{
redirectTau(target, ...)
}
\arguments{
\item{target}{The target folder or files.}

\item{...}{Passed on to \code{\link{redirect}}.}
}
\description{
redirectTau will be called by redirectSource when type = "Tau". Redirects
the Tau source folder to the target folder.
}
\examples{
\dontrun{
redirectSource("Tau", "a/different/tau-source-folder")
a <- readSource("Tau", "paper")
}
}
\author{
Pascal Sauer
}
