% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionscode.R
\name{regionscode}
\alias{regionscode}
\title{Tool: regionscode}
\usage{
regionscode(mapping)
}
\arguments{
\item{mapping}{Either a path to a mapping or an already read-in mapping as
data.frame.}
}
\value{
A md5-based regionscode which describes the given mapping
}
\description{
Given a regionmapping (mapping between ISO countries and regions) the
function calculates a regionscode which is basically the md5sum of a reduced
form of the mapping. The regionscode is unique for each regionmapping and
can be used to clearly identify a given regionmapping. In addition several
checks are performed to make sure that the given input is a proper
regionmapping
}
\examples{

\dontrun{
regionscode("regionmappingMAgPIE.csv")
}

}
\seealso{
\code{\link{fingerprint}}, \code{\link[digest]{digest}}
}
\author{
Jan Philipp Dietrich
}
