% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoryProfiling.R
\name{reportMemoryProfiling}
\alias{reportMemoryProfiling}
\title{reportMemoryProfiling}
\usage{
reportMemoryProfiling(start, callString)
}
\arguments{
\item{start}{the baseline as returned by \code{\link{startMemoryProfiling}};
if NULL nothing is reported}

\item{callString}{the function call the measurement belongs to}
}
\description{
Ends a memory profiling block started by \code{\link{startMemoryProfiling}} and
writes the collected numbers to the log.
}
\seealso{
\code{\link{startMemoryProfiling}}
}
\author{
Patrick Rein
}
\keyword{internal}
