% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTerraOptions.R
\name{setTerraOptions}
\alias{setTerraOptions}
\title{setTerraOptions}
\usage{
setTerraOptions()
}
\value{
Invisibly, the list of terra options
}
\description{
Uses `terra::terraOptions` to temporarily (until the current scope ends) set the terra
tempdir to a temporary folder in `getConfig("tmpfolder")`.
This way temp files are written ot the tempfolder specified by the madrat config. Also
`todisk` is set to TRUE and `memfrac` to 0.5, these are based on experience and reduce memory problems.
}
\author{
Pascal Führlich
}
