% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolConditionalReplace.R
\name{toolConditionalReplace}
\alias{toolConditionalReplace}
\title{toolConditionalReplace}
\usage{
toolConditionalReplace(x, conditions, replaceby = 0)
}
\arguments{
\item{x}{magpie object on cellular level}

\item{conditions}{vector of conditions for values, that should be removed e.g. "== NA", "< 0" (order matters)}

\item{replaceby}{value which should be used instead (can be a vector of same length as conditions as well)}
}
\value{
return changed input data
}
\description{
Sets values (NA, negative, ..) to zero
}
\author{
Kristine Karstens
}
