% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCountry2isocode.R
\name{toolCountry2isocode}
\alias{toolCountry2isocode}
\title{toolCountry2isocode}
\usage{
toolCountry2isocode(
  country,
  warn = TRUE,
  ignoreCountries = NULL,
  type = NULL,
  mapping = NULL
)
}
\arguments{
\item{country}{A vector of country names}

\item{warn}{whether warnings should be printed now or in the end of the
whole process as notes}

\item{ignoreCountries}{A vector of country names/codes that exist in the
data and that should be removed but without creating a
warning (they will be removed in any case). You should use that argument if
you are certain that the given entries should be actually removed from the
data.}

\item{type}{deprecated and will be removed soon!}

\item{mapping}{additional mappings as a names vector}
}
\value{
the ISO 3166-1 alpha 3 country code
}
\description{
Function used to convert country names from the long name to the ISO 3166-1
alpha 3 country code
}
\examples{

toolCountry2isocode("Germany")
toolCountry2isocode(c("Germany","Fantasyland"),mapping=c("Fantasyland"="BLA"))
}
\seealso{
\code{\link{readSource}},\code{\link{getSources}}
}
\author{
Jan Philipp Dietrich, Anastasis Giannousakis
}
