% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolFillYears.R
\name{toolFillYears}
\alias{toolFillYears}
\title{toolFillYears}
\usage{
toolFillYears(x, years)
}
\arguments{
\item{x}{MAgPIE object to be continued.}

\item{years}{vector of years as digits or in mag year format}
}
\value{
MAgPIE object with completed time dimensionality.
}
\description{
Inter- and extrapolates a historical dataset for a given time period.
}
\author{
Kristine Karstens
}
