% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolMappingFile.R
\name{toolMappingFile}
\alias{toolMappingFile}
\title{Tool: MappingFile}
\usage{
toolMappingFile(
  type,
  name,
  readcsv = FALSE,
  error.missing = TRUE,
  where = "mappingfolder"
)
}
\arguments{
\item{type}{Mapping type ("regional", or "sectoral")}

\item{name}{File name of the mapping file.}

\item{readcsv}{if true, file read in}

\item{error.missing}{Boolean which decides whether an error is returned if
the mapping file does not exist or not.}

\item{where}{location to look for the mapping, either "mappingfolder" or
the name of a package which contains the mapping}
}
\value{
The complete path to the mapping file.
}
\description{
Functions which calculates the path to a mapping file
}
\examples{

\dontrun{toolMappingFile("sectoral","structuremappingFE.csv")}


}
\seealso{
\code{\link{calcOutput}}
}
\author{
Jan Philipp Dietrich
}
