% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolNAreplace.R
\name{toolNAreplace}
\alias{toolNAreplace}
\title{Tool: NA replace}
\usage{
toolNAreplace(x, weight, replaceby = 0)
}
\arguments{
\item{x}{data}

\item{weight}{aggregation weight}

\item{replaceby}{value which should be used instead of NA}
}
\value{
a list containing x and weight
}
\description{
Functions removes NAs in x and weight
}
\seealso{
\code{\link{calcOutput}}
}
\author{
Benjamin Bodirsky
}
