% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolOrderCells.R
\name{toolOrderCells}
\alias{toolOrderCells}
\title{toolOrderCells}
\usage{
toolOrderCells(x, dim = 1.2, na.rm = FALSE)
}
\arguments{
\item{x}{magclass object that should be ordered}

\item{dim}{subdimension which contains the cell information}

\item{na.rm}{boolean deciding how to deal with non-integer information
in cellular column. If FALSE, non-integer values will lead to a return
of the unsorted object, if TRUE non-integer cells will be removed from
the data set and the rest will get sorted}
}
\value{
ordered data in magclass format
}
\description{
reorder numbered spatial units (cells, clusters) by number. Function
will return the unmodified object, if the given subdimension does not exist
or does not contain cell information.
}
\author{
Kristine Karstens, Jan Philipp Dietrich
}
