% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolstartmessage.R
\name{toolstartmessage}
\alias{toolstartmessage}
\title{Tool: Start message}
\usage{
toolstartmessage(level = NULL)
}
\arguments{
\item{level}{This argument allows to establish a hierarchy of print
statements. The hierarchy is preserved for the next vcat executions.
Currently this setting can have 4 states: NULL (nothing will be changed), 0
(reset hierarchies), "+" (increase hierarchy level by 1) and "-" (decrease
hierarchy level by 1).}
}
\value{
a list containing diagnostic information required by \code{\link{toolendmessage}}
}
\description{
Function writes a process start message and performs some diagnostics
}
\examples{

\dontrun{
tmp <- function(bla=NULL) {
  startinfo <- toolstartmessage("+")
  print(bla)
  toolendmessage(startinfo,"-")
  }
tmp(bla=99)
}

}
\seealso{
\code{\link{toolendmessage}}, \code{\link{vcat}}
}
\author{
Jan Philipp Dietrich
}
