% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolXlargest.R
\name{toolXlargest}
\alias{toolXlargest}
\title{toolXlargest}
\usage{
toolXlargest(type, range = 1:20, years = NULL, elements = NULL, ...)
}
\arguments{
\item{type}{calcOutput function that shall be used for ranking}

\item{range}{the position of the countries in the top X which should be returned.}

\item{years}{range of years that shall be summed for ranking.  If NULL, the sum of all years is used.}

\item{elements}{range of elements that shall be summed for ranking. If NULL, all elements are used.}

\item{...}{further parameters will be handed on to calcOutput function type.}
}
\value{
vector with ISO country codes
}
\description{
Selects the largest countries of a country-level magpie object
}
\examples{

\dontrun{ 
top10 <- toolXlargest(type="TauTotal",range=1:10)
}

}
\author{
Benjamin Leon Bodirsky, Jan Philipp Dietrich
}
