% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcat.R
\name{vcat}
\alias{vcat}
\title{Tool: Verbosity Cat}
\usage{
vcat(verbosity, ..., level = NULL)
}
\arguments{
\item{verbosity}{The lowest verbosity level for which this message should be
shown (verbosity = -1 means no information at all, 0 = only warnings, 1 =
warnings and execution informations, 2 = full information). If the verbosity
is set to 0 the message is written as warning, if the verbosity is set
higher than 0 it is written as a normal cat message.}

\item{...}{The message to be shown}

\item{level}{This argument allows to establish a hierarchy of print
statements. The hierarchy is preserved for the next vcat executions.
Currently this setting can have 4 states: NULL (nothing will be changed), 0
(reset hierarchies), "+" (increase hierarchy level by 1) and "-" (decrease
hierarchy level by 1).}
}
\description{
Function which returns information based on the verbosity setting
}
\examples{

\dontrun{
vcat(2,"Hello world!")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Jan Philipp Dietrich
}
