% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeDependencies.R
\name{visualizeDependencies}
\alias{visualizeDependencies}
\title{visualizeDependencies}
\usage{
visualizeDependencies(
  ...,
  direction = "both",
  order = 2,
  filter = NULL,
  packages = getConfig("packages"),
  filename = NULL
)
}
\arguments{
\item{...}{function(s) to be analyzed}

\item{direction}{Character string, either “in”, “out” or "both". If “in” all sources
feeding into the function are listed. If “out” consumer of the function are listed.
If “both” the union of "in" and "out" is returned.}

\item{order}{order of dependencies. Order 1 would be only functions directly called from (in case
of direction "in") or directly calling (in case of direction "out") are shown. Order 2 will also
show direct dependencies of the order 1 dependencies, order 3 also the direct dependencies from
order 2 dependencies, etc.}

\item{filter}{regular expression to describe elements which should be excluded from visualization
(e.g. "^tool" to exclude all tool functions)}

\item{packages}{packages to use when searching dependencies}

\item{filename}{If a filename is provided, the resulting graph will be saved}
}
\description{
Creates a graphical visualization of dependencies between functions in the mr-universe.
}
\seealso{
\code{\link{getDependencies}}, \code{\link{getMadratGraph}}, \code{\link{getMadratInfo}}
}
\author{
Debbora Leip, Jan Philipp Dietrich
}
