% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withMadratLogging.R
\name{withMadratLogging}
\alias{withMadratLogging}
\title{Tool: withMadratLogging}
\usage{
withMadratLogging(expr, logOnly = TRUE, warningsAsErrors = FALSE)
}
\arguments{
\item{expr}{expression to be evaluated.}

\item{logOnly}{passed to vcat, determines if warning/error is thrown after logging.
If omitted in a nested call, the enclosing logging policy is inherited.}

\item{warningsAsErrors}{whether warnings should be logged and handled as errors.
If omitted in a nested call, the enclosing logging policy is inherited.}
}
\description{
Function will activate madrat logging facilities for all code provided
to this function. This means that \code{message}, \code{warning} and
\code{stop} calls will also report to the madrat log output
}
\examples{
\dontrun{
madrat:::withMadratLogging(message("Hello world!"))
}

}
\seealso{
\code{\link{vcat}}
}
\author{
Jan Philipp Dietrich
}
\keyword{internal}
