% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappedCalcOutput.R
\name{wrappedCalcOutput}
\alias{wrappedCalcOutput}
\title{wrappedCalcOutput}
\usage{
wrappedCalcOutput(
  calc,
  aggregate = TRUE,
  file = NULL,
  years = NULL,
  round = NULL,
  signif = NULL,
  supplementary = FALSE,
  append = FALSE,
  warnNA = TRUE,
  na_warning = NULL,
  try = FALSE,
  regionmapping = NULL,
  writeArgs = NULL,
  ...
)
}
\arguments{
\item{calc}{A function to calculate the data. Mandatory. Function name must start with 'calc'.}

\item{aggregate}{Boolean indicating whether output data aggregation should be performed or
not, "GLO" (or "glo") for aggregation to one global region,
"REG+GLO" (or "regglo") for a combination of regional and global data.}

\item{file}{A file name. If given the output is written to that file in the
outputfolder as specified in the config.}

\item{years}{A vector of years that should be returned. If set to NULL all
available years are returned.}

\item{round}{Number of decimal places to round to.  Ignored if \code{NULL}.  See
\code{\link[=round]{round()}} for details.}

\item{signif}{Number of significant digits to round to.  Ignored if \code{NULL}.
See \code{\link[=signif]{signif()}} for details.}

\item{supplementary}{boolean deciding whether supplementary information such as weight should be
returned or not. If set to TRUE a list of elements will be returned!}

\item{append}{boolean deciding whether the output data should be appended in the existing file.
Works only when a file name is given in the function call.}

\item{warnNA}{boolean deciding whether NAs in the data set should create a warning or not}

\item{na_warning}{deprecated, please use \code{warnNA} instead}

\item{try}{if set to TRUE the calculation will only be tried and the script will continue even if
the underlying calculation failed. If set to FALSE calculation will stop with an error in such a
case. This setting will be overwritten by the global setting debug=TRUE, in which try will be
always interpreted as TRUE.}

\item{regionmapping}{alternative regionmapping to use for the given calculation. It will temporarily
overwrite the global setting just for this calculation.}

\item{writeArgs}{a list of additional, named arguments to be supplied to
the corresponding write function}

\item{...}{Additional settings directly forwarded to the corresponding
calculation function}
}
\value{
magpie object with the requested output data either on country or on
regional level depending on the choice of argument "aggregate" or a list of information
if supplementary is set to TRUE.
}
\description{
Minimal wrapper for calcOutput.
Beware that no checks are done that the input follows the required naming conventions.
}
\author{
Sally Dacie, Pascal Sauer
}
