% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappedReadSource.R
\name{wrappedReadSource}
\alias{wrappedReadSource}
\title{wrappedReadSource}
\usage{
wrappedReadSource(sourceObject, subtype = NULL, subset = NULL)
}
\arguments{
\item{sourceObject}{A list with the following elements:
\itemize{
  \item read A function to read the data. Mandatory.
  \item correct A function to correct the data. Optional.
  \item covert A function to convert the data (to country level). Optional.
  \item download A function to download the data. Optional.
}}

\item{subtype}{A character string. For some sources there are subtypes of the source, for these
sources the subtype can be specified with this argument. If a source does not
have subtypes, subtypes should not be set.}

\item{subset}{A character string. Similar to \code{subtype} a source can also have \code{subsets}.
A \code{subsets} can be used to only read part of the data. This can in particular make sense
for huge data sets where reading in the whole data set might be impractical or even infeasible.}
}
\value{
The read-in data, usually a magpie object. If supplementary is TRUE a list including
the data and metadata is returned instead. The temporal and data dimensionality
should match the source data. The spatial dimension should either match the source data or,
if the convert argument is set to TRUE, should be on ISO code country level.
}
\description{
Minimal wrapper for readSource.
Beware that no checks are done that the input follows the required naming conventions.
}
\author{
Pascal Sauer
}
