% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMapping.R
\name{addMapping}
\alias{addMapping}
\title{addMapping}
\usage{
addMapping(filename, mapping = NULL)
}
\arguments{
\item{filename}{The name of the the region mapping that should added including
file ending (e.g. "regionmappingREMIND.csv"). Supported formats are currently
".csv" and ".rds".}

\item{mapping}{Mapping provided as data.frame, or NULL. If a mapping is
provided the data will be written in the mapping file of the given file
(potentially replacing existing data). If NULL the mapping from the given
file is used.}
}
\description{
Function whichs adds another mapping to the current list of extramappings
in the madrat configuration (see \code{\link{setConfig}}) and stores
the mapping in the mapping folder as well as output folder.
}
\examples{
\dontrun{
addMapping("regionmappingH12.csv")
}
}
\seealso{
\code{\link{setConfig}}
}
\author{
Jan Philipp Dietrich
}
