% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheCleanup.R
\name{cacheCleanup}
\alias{cacheCleanup}
\title{cacheCleanup}
\usage{
cacheCleanup(
  lifespanDays,
  cacheFolder,
  ask = TRUE,
  readlineFunction = readline
)
}
\arguments{
\item{lifespanDays}{Files older than this many days are deleted.}

\item{cacheFolder}{Path to the folder to delete files from.}

\item{ask}{If TRUE (the default) the user is asked before deleting.}

\item{readlineFunction}{A function to get input from the user. Only intended for testing.}
}
\description{
Delete cache files older than a specified period of time, based on atime file metadata.
}
