% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheFormat.R
\name{cacheFormat}
\alias{cacheFormat}
\title{cacheFormat}
\usage{
cacheFormat(name = getConfig("cacheformat"))
}
\arguments{
\item{name}{Name of the format, defaults to the currently configured one.}
}
\value{
The format definition, with the format name added as element "name".
}
\description{
Look up a registered cache format definition.
}
\keyword{internal}
