% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheGet.R
\name{cacheGet}
\alias{cacheGet}
\title{Tool: cacheGet}
\usage{
cacheGet(prefix, type, args = NULL)
}
\arguments{
\item{prefix}{function prefix (e.g. "calc" or "read")}

\item{type}{output type (e.g. "TauTotal")}

\item{args}{a list of named arguments used to call the given function}
}
\value{
cached data if available, otherwise NA
attr(, "id") will be set to the cache file name that should be written if the
data has to be recalculated.
attr(, "readFile") will be set to the cache file that was successfully read,
and is absent otherwise. This can differ from attr(, "id") if the data was
read from an rds file while another cache format is configured.
}
\description{
Load fitting cache data (if available)
}
\examples{
madrat:::cacheGet("calc", "TauTotal")
}
\seealso{
\code{\link{cachePut}}, \code{\link{cacheName}}
}
\author{
Jan Philipp Dietrich, Pascal Sauer
}
\keyword{internal}
