% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheGet.R
\name{cacheGet}
\alias{cacheGet}
\title{Tool: cacheGet}
\usage{
cacheGet(prefix, type, args = NULL, graph = NULL, ...)
}
\arguments{
\item{prefix}{function prefix (e.g. "calc" or "read")}

\item{type}{output type (e.g. "TauTotal")}

\item{args}{a list of named arguments used to call the given function}

\item{graph}{A madrat graph as returned by \code{\link{getMadratGraph}}.
Will be created with \code{\link{getMadratGraph}} if not provided.}

\item{...}{Additional arguments for \code{\link{getMadratGraph}} in case
that no graph is provided (otherwise ignored)}
}
\value{
cached data, if cache is available, otherwise NULL
}
\description{
Load fitting cache data (if available)
}
\examples{
madrat:::cacheGet("calc", "TauTotal", packages = "madrat")
}
\seealso{
\code{\link{cachePut}}, \code{\link{cacheName}}
}
\author{
Jan Philipp Dietrich
}
