% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheName.R
\name{cacheName}
\alias{cacheName}
\title{cacheName}
\usage{
cacheName(prefix, type, args = NULL)
}
\arguments{
\item{prefix}{function prefix (e.g. "calc" or "read")}

\item{type}{output type (e.g. "TauTotal")}

\item{args}{a list of named arguments used to call the given function}
}
\value{
absolute path to a cache file that does not necessarily exist
}
\description{
Get the name of a cache file corresponding to the given args
}
\note{
With \code{setConfig(forcecache=TRUE)} cacheName will also return
cache file names with deviating fingerprint if no fitting cache file is found
(if there are multiple it will just return the newest one).
}
\examples{
madrat:::cacheName("calc", "TauTotal")
}
\seealso{
\code{\link{cachePut}}
}
\author{
Jan Philipp Dietrich, Pascal Sauer
}
