% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cachePut.R
\name{cachePut}
\alias{cachePut}
\title{cachePut}
\usage{
cachePut(x, prefix, type, fname, callString)
}
\arguments{
\item{x}{data that should be written to cache}

\item{prefix}{function prefix (e.g. "calc" or "read")}

\item{type}{output type (e.g. "TauTotal")}

\item{fname}{The name of the cache file to be written.}

\item{callString}{A string representation of the function call that leads
to the cache file being written. Will be attached as an attribute.}
}
\description{
Save data to cache
}
\examples{
\dontrun{
fname <- madrat:::cacheName("calc", "Example")
madrat:::cachePut(1, "calc", "Example", fname, 'calcOutput("Example")')
}
}
\seealso{
\code{\link{cachePut}}, \code{\link{cacheName}}

Other cache management: 
\code{\link{cacheArgumentsHash}()},
\code{\link{cacheCleanup}()},
\code{\link{cacheCopy}()},
\code{\link{cacheGet}()},
\code{\link{cacheName}()},
\code{\link{getNonDefaultArguments}()}
}
\author{
Jan Philipp Dietrich, Pascal Sauer
}
\concept{cache management}
