% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cachePut.R
\name{cachePut}
\alias{cachePut}
\title{cachePut}
\usage{
cachePut(x, prefix, type, args, callString)
}
\arguments{
\item{x}{data that should be written to cache}

\item{prefix}{function prefix (e.g. "calc" or "read")}

\item{type}{output type (e.g. "TauTotal")}

\item{args}{a list of named arguments used to call the given function}

\item{callString}{A string representation of the function call that leads
to the cache file being written. Will be attached as an attribute.}
}
\description{
Save data to cache
}
\examples{
\dontrun{
madrat:::cachePut(1, "calc", "Example", NULL, 'calcOutput("Example")')
}
}
\seealso{
\code{\link{cachePut}}, \code{\link{cacheName}}
}
\author{
Jan Philipp Dietrich
}
