% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cachePut.R
\name{cachePut}
\alias{cachePut}
\title{Tool: cachePut}
\usage{
cachePut(x, prefix, type, args = NULL, graph = NULL, ...)
}
\arguments{
\item{x}{data that should be written to cache}

\item{prefix}{function prefix (e.g. "calc" or "read")}

\item{type}{output type (e.g. "TauTotal")}

\item{args}{a list of named arguments used to call the given function}

\item{graph}{A madrat graph as returned by \code{\link{getMadratGraph}}
Will be created with \code{\link{getMadratGraph}} if not provided.}

\item{...}{Additional arguments for \code{\link{getMadratGraph}} in case
that no graph is provided (otherwise ignored)}
}
\description{
Save data to cache
}
\examples{
\dontrun{
example <- 1
madrat:::cachePut(example, "calc", "Example", packages = "madrat")
}
}
\seealso{
\code{\link{cachePut}}, \code{\link{cacheName}}
}
\author{
Jan Philipp Dietrich
}
