% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheFormat.R
\name{cacheRead}
\alias{cacheRead}
\alias{cacheWrite}
\alias{cacheToRds}
\title{cacheRead / cacheWrite / cacheToRds}
\usage{
cacheRead(file)

cacheWrite(x, file, extension = file_ext(file))

cacheToRds(input, output)
}
\arguments{
\item{file}{Path of the cache file to be read/written.}

\item{x}{Object to be written.}

\item{extension}{Format extension to use. Defaults to the extension of
\code{file}, but has to be given explicitly if \code{file} does not end on it
(e.g. when writing to a temporary file name).}

\item{input}{Path of the cache file to be converted.}

\item{output}{Path of the rds file to be created.}
}
\description{
Read, write and convert cache files using the format belonging to their file
extension. Dispatching on the extension rather than on the configured format
is what allows madrat to still read rds cache files while writing a different
format (see \code{\link{cacheName}}).
}
\section{Functions}{
\itemize{
\item \code{cacheWrite()}: write a cache file

\item \code{cacheToRds()}: convert a cache file to rds

}}
\author{
Patrick Rein
}
\keyword{internal}
