% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTauTotal.R
\name{calcTauTotal}
\alias{calcTauTotal}
\title{Calculate total tau}
\usage{
calcTauTotal(source = "paper")
}
\arguments{
\item{source}{data source, either "paper" (default) or "historical".}
}
\value{
Total tau data and corresponding weights as a list of two MAgPIE
objects
}
\description{
This function prepares total tau values for use. As the source data already
provides all required information this function purely removes not required
data and moves xref values to the weighting object which is required for
aggregation.
}
\examples{
\dontrun{
calcOutput("TauTotal")
}
}
\seealso{
\code{\link{calcOutput}}, \code{\link{readTau}},
\code{\link{convertTau}}
}
\author{
Jan Philipp Dietrich
}
