% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareData.R
\name{compareData}
\alias{compareData}
\title{compareData}
\usage{
compareData(x, y, tolerance = 10^-5, yearLim = NULL)
}
\arguments{
\item{x}{Either a tgz file or a folder containing data sets}

\item{y}{Either a tgz file or a folder containing data sets}

\item{tolerance}{tolerance level below which differences will
get ignored}

\item{yearLim}{year until when the comparison should be performed.
Useful to check if data is identical until a certain year.}
}
\description{
Compares the content of two data archives and looks for similarities and differences
}
\seealso{
\code{\link{setConfig}}, \code{\link{calcTauTotal}},
}
\author{
Jan Philipp Dietrich, Florian Humpenoeder
}
