% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBottlenecks.R
\name{findBottlenecks}
\alias{findBottlenecks}
\title{findBottlenecks}
\usage{
findBottlenecks(file, unit = "min", cumulative = TRUE)
}
\arguments{
\item{file}{path to a log file or content of a log as character vector}

\item{unit}{unit for runtime information, either "s" (seconds), "min" (minutes) or "h" (hours)}

\item{cumulative}{boolean deciding whether calls to the same function should be aggregated or not}
}
\value{
A data.frame sorted by net runtime showing for the different data processing functions their
total runtime "time" (including the execution of all sub-functions) and net runtime "net" (excluding
the runtime of sub-functions) and their share of total runtime.
}
\description{
Analyzes a log from a retrieveData run, extracts runtime information for all called functions
and identifies most critical bottlenecks.
}
\author{
Jan Philipp Dietrich
}
