% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOldFiles.R
\name{findOldFiles}
\alias{findOldFiles}
\title{findOldFiles}
\usage{
findOldFiles(daysThreshold, path)
}
\arguments{
\item{daysThreshold}{Files older than this many days are returned.}

\item{path}{Path to where to look for old files.}
}
\description{
Find files older than the specified number of days, based on atime/last access time file metadata. This
metadata is not always available and it is system dependent what atime/last access time actually means.
}
