% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullEXAMPLE.R
\name{fullEXAMPLE}
\alias{fullEXAMPLE}
\title{fullExample}
\usage{
fullEXAMPLE(rev = 0, dev = "", extra = "Example argument")
}
\arguments{
\item{rev}{data revision which should be used/produced. Will be converted to
\code{\link[base]{numeric_version}} when called via \code{\link{retrieveData}}.}

\item{dev}{development suffix to distinguish development versions for the same data
revision. This can be useful to distinguish parallel lines of development.}

\item{extra}{additional argument which - when changed - does not require a re-computation
of the portable unaggegrated collection (puc) file.
\code{\link{setConfig}} (e.g. for setting the mainfolder if not already set
properly).}
}
\description{
Example for class of fullX functions. Can be used as template for a new function
or for testing the basic functionality
}
\examples{
\dontrun{
retrieveData("example", rev = "2.1.2", dev = "test", regionmapping = "regionmappingH12.csv")
}

}
\seealso{
\code{\link{readSource}},\code{\link{getCalculations}},\code{\link{calcOutput}},\code{\link{setConfig}}
}
\author{
Jan Philipp Dietrich
}
