% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCode.R
\name{getCode}
\alias{getCode}
\title{getCode}
\usage{
getCode(
  packages = installedMadratUniverse(),
  globalenv = getConfig("globalenv")
)
}
\arguments{
\item{packages}{A character vector with packages for which the available Sources/Calculations should be returned}

\item{globalenv}{Boolean deciding whether sources/calculations in the global environment should be included or not}
}
\value{
A named vector with condensed function code
}
\description{
Extract function code from madrat-style functions in specified packages
}
\seealso{
\code{\link{getMadratGraph}}
}
\author{
Jan Philipp Dietrich
}
