% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLinkFunction.R
\name{getLinkFunction}
\alias{getLinkFunction}
\title{getLinkFunction}
\usage{
getLinkFunction()
}
\value{
A function with arguments "from" and "to" which should behave like
file.symlink on all platforms.
}
\description{
Returns a function that creates a symlink, hardlink, junction, or copy of
files and directories, depending on OS capabilities (usually symlinks are
not supported on Windows).
}
\author{
Pascal Sauer
}
