% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSourceFolder.R
\name{getSourceFolder}
\alias{getSourceFolder}
\title{getSourceFolder}
\usage{
getSourceFolder(type, subtype)
}
\arguments{
\item{type}{Dataset name, e.g. "Tau" for \code{\link{readTau}}}

\item{subtype}{Subtype of the dataset, e.g. "paper" for \code{\link{readTau}}, NULL is allowed}
}
\value{
Path to source data files
}
\description{
Return the path to source data files for the given type and subtype. This
applies redirections, see \code{\link{redirectSource}} for more details.
}
\author{
Pascal Sauer
}
