% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeConfig.R
\name{initializeConfig}
\alias{initializeConfig}
\title{initializeConfig}
\usage{
initializeConfig(verbose = TRUE)
}
\arguments{
\item{verbose}{boolean deciding whether status information/updates should be shown or not}
}
\description{
Checks whether configuration already has been set. If not, it will be initialized
with default settings or (if available) system settings. All madrat folders (see
\code{\link{setConfig}} for documentation which folders are available) will be set
to the system environment variables MADRAT_SOURCEFOLDER, MADRAT_CACHEFOLDER, etc.
if they exist, NA otherwise. NA means subfolders of the mainfolder are used.
}
\seealso{
\code{\link{getMainfolder}}, \code{\link{getConfig}}, \code{\link{setConfig}}
}
\author{
Jan Philipp Dietrich
}
