% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isWrapperActive.R
\name{isWrapperActive}
\alias{isWrapperActive}
\alias{setWrapperActive}
\alias{setWrapperInactive}
\title{isWrapperActive}
\usage{
isWrapperActive(name)

setWrapperActive(name)

setWrapperInactive(name)
}
\arguments{
\item{name}{name of the wrapper in question (e.g. "calcOutput")}
}
\description{
Support functions which checks whether a given wrapper function is currently in-use
or not or which locally activate or deactivate a wrapper (setting will be
automatically resetted when a function finishes).
}
\section{Functions}{
\itemize{
\item \code{setWrapperActive()}: set wrapper activity status to on

\item \code{setWrapperInactive()}: set wrapper activity status to off

}}
\author{
Jan Philipp Dietrich
}
