% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localRedirect.R
\name{localRedirect}
\alias{localRedirect}
\title{localRedirect}
\usage{
localRedirect(type, subtype = NULL, ..., target)
}
\arguments{
\item{type}{Dataset name, e.g. "Tau" for \code{\link{readTau}}}

\item{subtype}{Not supported for now, must be NULL. This is here to simplify
a potential later implementation of subtype redirections.}

\item{...}{Additional arguments, passed on to source-specific inject function if it exists}

\item{target}{Path to the new source folder, NULL to remove the redirection}
}
\value{
Invisibly, a list of all redirections where names are types and
values are the paths these types are redirected to.
}
\description{
Redirect a given dataset type to a different source folder. The redirection
is local, so it will be reset when the current function call returns. See
example for more details.
}
\examples{
\dontrun{
f <- function() {
  localRedirect("Tau", target = "~/TauExperiment")
  # the following call will change directory
  # into ~/TauExperiment instead of <getConfig("sourcefolder")>/Tau
  readSource("Tau")
}
f()
# Tau is only redirected in the local environment of f,
# so it will use the usual source folder here
readSource("Tau")
}
}
\author{
Pascal Sauer
}
