% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madapply.R
\name{madapply}
\alias{madapply}
\title{madapply}
\usage{
madapply(...)
}
\arguments{
\item{...}{placeholder}
}
\description{
This function is defunct and will be completely removed soon.
}
\author{
Jan Philipp Dietrich
}
