% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/madlapply.R
\name{madlapply}
\alias{madlapply}
\title{madlapply}
\usage{
madlapply(...)
}
\arguments{
\item{...}{placeholder}
}
\description{
This function is defunct and will be completely removed soon.
}
\author{
Jan Philipp Dietrich
}
