% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepExtendedComment.R
\name{prepExtendedComment}
\alias{prepExtendedComment}
\title{prepExtendedComment}
\usage{
prepExtendedComment(x, type, functionCallString, warn = TRUE)
}
\arguments{
\item{x}{list containing the metadata to be condensed}

\item{type}{output type, e.g. "TauTotal"}

\item{functionCallString}{A string representation of the function call
that created the object this comment is attached to}

\item{warn}{boolean indicating whether warnings should be triggered
if entries are missing, or not.}
}
\description{
Helper function condense metadata information into
an extended comment entry
}
\author{
Jan Philipp Dietrich, Pascal Sauer
}
