% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redirectSource.R
\name{redirectSource}
\alias{redirectSource}
\title{redirectSource}
\usage{
redirectSource(type, target, ..., linkOthers = TRUE, local = TRUE)
}
\arguments{
\item{type}{The source dataset type. Passed on to the specific redirect
function or \code{\link{redirect}}.}

\item{target}{The target folder or files. Passed on to the specific redirect
function or \code{\link{redirect}}.}

\item{...}{Additional arguments, passed on to the specific redirect function.}

\item{linkOthers}{Passed on to the specific redirect function or \code{\link{redirect}}.}

\item{local}{Passed on to the specific redirect function or \code{\link{redirect}}.}
}
\value{
The result of the specific redirect function or \code{\link{redirect}}.
}
\description{
redirectSource will call a source specific redirect function if it exists
(called e.g. redirectTau), in which case the arguments are passed on to that
function. If such a function is not available \code{\link{redirect}} is called.
}
\author{
Pascal Sauer
}
