% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveData.R
\name{retrieveData}
\alias{retrieveData}
\title{retrieveData}
\usage{
retrieveData(model, rev = 0, dev = "", cachetype = "rev", ...)
}
\arguments{
\item{model}{The names of the model for which the data should be provided
(e.g. "magpie").}

\item{rev}{data revision which should be used/produced. Format must be compatible to
\code{\link[base]{numeric_version}}.}

\item{dev}{development suffix to distinguish development versions for the same data
revision. This can be useful to distinguish parallel lines of development.}

\item{cachetype}{defines what cache should be used. "rev" points to a cache
shared by all calculations for the given revision, "def" points to the cache
as defined in the current settings and "tmp" temporarily creates a cache
folder for the calculations and deletes it afterwards again}

\item{...}{(Optional) Settings that should be changed using
\code{\link{setConfig}} (e.g. regionmapping).}
}
\description{
Function to retrieve a predefined collection of calculations for a specific
regionmapping.
}
\examples{

\dontrun{ 
retrieveData("example", rev="2.1.1", dev="test", regionmapping="regionmappingH12.csv")
}

}
\seealso{
\code{\link{calcOutput}},\code{\link{setConfig}}
}
\author{
Jan Philipp Dietrich, Lavinia Baumstark
}
