% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setConfig.R
\name{setConfig}
\alias{setConfig}
\alias{localConfig}
\title{setConfig}
\usage{
setConfig(
  regionmapping = NULL,
  extramappings = NULL,
  packages = NULL,
  globalenv = NULL,
  enablecache = NULL,
  verbosity = NULL,
  mainfolder = NULL,
  sourcefolder = NULL,
  cachefolder = NULL,
  mappingfolder = NULL,
  outputfolder = NULL,
  pucfolder = NULL,
  tmpfolder = NULL,
  nolabels = NULL,
  forcecache = NULL,
  ignorecache = NULL,
  cachecompression = NULL,
  hash = NULL,
  diagnostics = NULL,
  debug = NULL,
  maxLengthLogMessage = NULL,
  redirections = NULL,
  .cfgchecks = TRUE,
  .verbose = TRUE,
  .local = FALSE
)

localConfig(...)
}
\arguments{
\item{regionmapping}{The name of the csv file containing the region mapping
that should be used for aggregation (e.g. "regionmappingREMIND.csv").}

\item{extramappings}{Names of additional mappings supplementing the given
region mapping. This allows for additional aggregation levels such as
subnational aggregation.}

\item{packages}{A character vector with packages in which corresponding
read and calc functions should be searched for}

\item{globalenv}{Boolean deciding whether sources/calculations in the global
environment should be included or not}

\item{enablecache}{Is deprecated and will be ignored. Please use
\code{ignorecache} instead.}

\item{verbosity}{an integer value describing the verbosity of the functions
(2 = full information, 1 = only warnings and execution information, 0 = only
warnings, -1 = no information)}

\item{mainfolder}{The mainfolder where all data can be found and should be
written to.}

\item{sourcefolder}{The folder in which all source data is stored (in
sub-folders with the name of the source as folder name). In the default case
this argument is set to NA meaning that the default folder should be used
which is <mainfolder>/sources}

\item{cachefolder}{The folder in which all cache files should be written to.
In the default case this argument is set to NA meaning that the default
folder should be used which is <mainfolder>/cache}

\item{mappingfolder}{A folder containing all kinds of mappings (spatial,
temporal or sectoral). In the default case this argument is set to NA
meaning that the default folder should be used which is
<mainfolder>/mappings}

\item{outputfolder}{The folder all outputs should be written to. In the
default case this argument is set to NA meaning that the default folder
should be used which is <mainfolder>/output}

\item{pucfolder}{The path where portable unaggregated collection (puc) files
are located. NA by default, which means <mainfolder>/puc}

\item{tmpfolder}{Path to a temp folder for temporary storage of files. By default
set to <mainfolder>/tmp}

\item{nolabels}{vector of retrieve models (e.g. "EXAMPLE" in case of "fullEXAMPLE")
which should NOT apply a replacement of known hashes with given code labels}

\item{forcecache}{Argument that allows to force madrat to read data from
cache if the corresponding cache files exist. It is either a boolean to
fully activate or deactivate the forcing or a vector of files (e.g. readTau, calcTauTotal)
or type (e.g. Tau, TauTotal) that should be read from cache in any case.}

\item{ignorecache}{Argument that allows madrat to ignore the forcecache argument for the
given vector of files (e.g. readTau, calcTauTotal) or types
(e.g. Tau, TauTotal) called by calcOutput or readSource.
The top level function must always be part of this list.}

\item{cachecompression}{logical or character string specifying whether cache files
use compression. TRUE corresponds to gzip compression, and character strings "gzip",
"bzip2" or "xz" specify the type of compression.}

\item{hash}{specifies the used hashing algorithm. Default is "xxhash32" and
all algorithms supported by \code{\link[digest]{digest}} can be used.}

\item{diagnostics}{Either FALSE (default) to avoid the creation of additional
log files or a file name for additional diagnostics information (without file ending).}

\item{debug}{Boolean which activates a debug mode. In debug mode all calculations will
be executed with try=TRUE so that calculations do not stop even if the previous calculation failed.
This can be helpful to get a full picture of errors rather than only seeing the first one. In addition
debug=TRUE will add the suffix "debug" to the files created to avoid there use in productive runs.
Furthermore, with debug=TRUE calculations will be rerun even if a corresponding tgz file
already exists.}

\item{maxLengthLogMessage}{in log messages evaluated arguments are printed if the resulting message
is shorter than this value, otherwise arguments are shown as passed, potentially with unevaluated variable names}

\item{redirections}{A list of source folder redirections, intended to be set
by \code{\link{redirectSource}}. See that function's documentation for more details.}

\item{.cfgchecks}{boolean deciding whether the given inputs to setConfig should be checked for
consistency or just be accepted (latter is only necessary in very rare cases and should not be used
in regular cases)}

\item{.verbose}{boolean deciding whether status information/updates should be shown or not}

\item{.local}{boolean deciding whether options are only changed until the end of the current function execution
OR environment for which the options should get changed.}

\item{...}{Arguments forwarded to setConfig}
}
\description{
This function manipulates the current madrat configuration.
In general, NULL means that the argument remains as it is whereas
all other inputs will overwrite the current setting.
For values which can be reset to NULL (currently only "extramappings")
you can achieve a reset by setting the value to "".
}
\section{Functions}{
\itemize{
\item \code{localConfig()}: A wrapper for setConfig(..., .local = TRUE)

}}
\note{
\code{setConfig} must only be used before the data processing is started and changes in the configuration
from within a download-, read-, correct-, convert-, calc-, or full-function are not allowed! Only allowed
configuration update is to add another \code{extramapping} via \code{\link{addMapping}}.
Currently the use of \code{setConfig} within any of these functions will trigger a warning, which is planned
to be converted into an error message in one of the next package updates!
}
\examples{
\dontrun{
setConfig(forcecache = c("readSSPall", "convertSSPall"))
}
}
\seealso{
\code{\link{getConfig}}, \code{\link{getISOlist}}
}
\author{
Jan Philipp Dietrich
}
