% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolendmessage.R
\name{toolendmessage}
\alias{toolendmessage}
\title{Tool: End message}
\usage{
toolendmessage(startdata, level = NULL)
}
\arguments{
\item{startdata}{a list containing diagnostic information provided by \code{\link{toolstartmessage}}}

\item{level}{This argument allows to establish a hierarchy of print
statements. The hierarchy is preserved for the next vcat executions.
Currently this setting can have 4 states: NULL (nothing will be changed), 0
(reset hierarchies), "+" (increase hierarchy level by 1) and "-" (decrease
hierarchy level by 1).}
}
\description{
This function writes a process end message and performs some diagnostics. It is always called after a corresponding
call to \code{\link{toolstartmessage}}.
}
\seealso{
\code{\link{toolstartmessage}}, \code{\link{vcat}}
}
\author{
Jan Philipp Dietrich
}
