% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolNAreplace.R
\name{toolNAreplace}
\alias{toolNAreplace}
\title{Tool: NA replace}
\usage{
toolNAreplace(x, weight = NULL, replaceby = 0, val.rm = NULL)
}
\arguments{
\item{x}{data}

\item{weight}{aggregation weight}

\item{replaceby}{value which should be used instead of NA. Either a single 
value or a MAgPIE object which can be expanded to the size of x (either same 
size or with lower dimensionality).}

\item{val.rm}{vector of values that should in addition be removed in x}
}
\value{
a list containing x and weight
}
\description{
Functions removes NAs, NaNs and infinite values in x and weight
}
\seealso{
\code{\link{calcOutput}}
}
\author{
Benjamin Bodirsky, Jan Philipp Dietrich
}
