% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolTimeSpline.R
\name{toolTimeSpline}
\alias{toolTimeSpline}
\title{toolTimeSpline}
\usage{
toolTimeSpline(x, dof = NULL)
}
\arguments{
\item{x}{magclass object that should be smoothed via a spline approximation}

\item{dof}{degrees of freedom per 100 years (similiar to an average range), is a proxy for the smoothness of the 
spline (smaller values = smoother)}
}
\value{
approximated data in magclass format
}
\description{
Smoothing a data set by replacing its values by its spline approximation using the given degrees of freedom.
}
\author{
Kristine Karstens, Felicitas Beier
}
