% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withMadratLogging.R
\name{withMadratLogging}
\alias{withMadratLogging}
\title{Tool: withMadratLogging}
\usage{
withMadratLogging(expr)
}
\arguments{
\item{expr}{expression to be evaluated.}
}
\description{
Function will activate madrat logging facilities for all code provided
to this function. This means that \code{message}, \code{warning} and
\code{stop} calls will also report to the madrat log output
}
\examples{
\dontrun{
madrat:::withMadratLogging(message("Hello world!"))
}

}
\seealso{
\code{\link{vcat}}
}
\author{
Jan Philipp Dietrich
}
