% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappedReadSource.R
\name{wrappedReadSource}
\alias{wrappedReadSource}
\title{wrappedReadSource}
\usage{
wrappedReadSource(
  read,
  correct = NULL,
  convert = NULL,
  subtype = NULL,
  subset = NULL
)
}
\arguments{
\item{read}{A function to read the data. Mandatory. Function name must start with 'read'. The
rest of the function name determines the 'type'.}

\item{correct}{A function to correct the data. Optional. Function name must start with 'correct'
and end with the type, same as for read.}

\item{subtype}{A character string. For some sources there are subtypes of the source, for these
sources the subtype can be specified with this argument. If a source does not
have subtypes, subtypes should not be set.}

\item{subset}{A character string. Similar to \code{subtype} a source can also have \code{subsets}.
A \code{subsets} can be used to only read part of the data. This can in particular make sense
for huge data sets where reading in the whole data set might be impractical or even infeasible.}

\item{covert}{A function to convert the data (putting it at country level). Optional. Function 
name must start with 'convert' and end with the type, same as for read.}
}
\description{
Minimal wrapper for readSource.
Beware that no checks are done that the input follows the required naming conventions.
}
